소켓 레스토랑 업데이트 수정

+++ src/common/entities/Lating.entity.ts
@@ -2,11 +2,11 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Column } from 'typeorm';
 
 export class Lating {
-  @Column({ type: 'float' })
+  @Column({ type: 'double' })
   @ApiProperty({ description: 'Lating_X', example: 'X 좌표입니다.' })
   x: number;
 
-  @Column({ type: 'float' })
+  @Column({ type: 'double' })
   @ApiProperty({ description: 'Lating_Y', example: 'Y 좌표입니다.' })
   y: number;
 }
